テーブル構造の修正
